Added loading UI and decomposed components
