+++ src/app/[pohid]/[chain]/[request]/loading.tsx
@@ -1,0 +1,3 @@
+import Loading from "components/Loading";
+
+export default Loading;
