+++ src/app/[pohid]/claim/page.tsx
@@ -7,7 +7,7 @@ import dynamic from "next/dynamic";
 import { redirect } from "next/navigation";
 import { machinifyId } from "utils/identifier";
 import { Hash } from "viem";
-import Loading from "app/[pohid]/loading";
+import Loading from "components/Loading";
 
 const Form = dynamic(() => import("./Form"), {
   ssr: false,
